Un-nest validator registration request
