+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/dsl/utils/Eth2BlockSigningRequestUtil.java
@@ -87,6 +87,7 @@ public class Eth2BlockSigningRequestUtil {
         signingRoot,
         forkInfo,
         null,
+        null,
         blockRequest,
         null,
         null,
@@ -110,6 +111,7 @@ public class Eth2BlockSigningRequestUtil {
         ArtifactType.BLOCK,
         signingRoot,
         forkInfo,
+        null,
         getBeaconBlock(),
         null,
         null,
