+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/tests/signing/Eth2DepositSigningAcceptanceTest.java
@@ -113,6 +113,7 @@ public class Eth2DepositSigningAcceptanceTest extends SigningAcceptanceTestBase 
             null,
             null,
             null,
+            null,
             depositMessage,
             null,
             null,
