+++ core/src/main/java/tech/pegasys/web3signer/core/service/http/handlers/signing/eth2/Eth2SigningRequestBody.java
@@ -17,6 +17,7 @@ import tech.pegasys.teku.api.schema.AttestationData;
 import tech.pegasys.teku.api.schema.BeaconBlock;
 import tech.pegasys.teku.api.schema.VoluntaryExit;
 import tech.pegasys.teku.api.schema.altair.ContributionAndProof;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.web3signer.core.service.http.ArtifactType;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -27,6 +28,8 @@ public class Eth2SigningRequestBody {
   private final ArtifactType type;
   private final Bytes signingRoot;
   private final ForkInfo fork_info;
+
+  private final UInt64 epoch;
   private final BeaconBlock beaconBlock; // phase 0
   private final BlockRequest blockRequest; // altair and onward
   private final AttestationData attestation;
@@ -39,13 +42,14 @@ public class Eth2SigningRequestBody {
   private final SyncAggregatorSelectionData syncAggregatorSelectionData;
   private final ContributionAndProof contributionAndProof;
 
-  private final ValidatorRegistrationRequest validatorRegistrationRequest;
+  private final ValidatorRegistration validatorRegistration;
 
   @JsonCreator
   public Eth2SigningRequestBody(
       @JsonProperty(value = "type", required = true) final ArtifactType type,
       @JsonProperty("signingRoot") final Bytes signingRoot,
       @JsonProperty("fork_info") final ForkInfo fork_info,
+      @JsonProperty("epoch") final UInt64 epoch,
       @JsonProperty("block") final BeaconBlock block,
       @JsonProperty("beacon_block") final BlockRequest blockRequest,
       @JsonProperty("attestation") final AttestationData attestation,
@@ -58,11 +62,11 @@ public class Eth2SigningRequestBody {
       @JsonProperty("sync_aggregator_selection_data")
           final SyncAggregatorSelectionData syncAggregatorSelectionData,
       @JsonProperty("contribution_and_proof") final ContributionAndProof contributionAndProof,
-      @JsonProperty("validator_registration_request")
-          final ValidatorRegistrationRequest validatorRegistrationRequest) {
+      @JsonProperty("validator_registration") final ValidatorRegistration validatorRegistration) {
     this.type = type;
     this.signingRoot = signingRoot;
     this.fork_info = fork_info;
+    this.epoch = epoch;
     this.beaconBlock = block;
     this.blockRequest = blockRequest;
     this.attestation = attestation;
@@ -74,7 +78,7 @@ public class Eth2SigningRequestBody {
     this.syncCommitteeMessage = syncCommitteeMessage;
     this.syncAggregatorSelectionData = syncAggregatorSelectionData;
     this.contributionAndProof = contributionAndProof;
-    this.validatorRegistrationRequest = validatorRegistrationRequest;
+    this.validatorRegistration = validatorRegistration;
   }
 
   @JsonProperty("type")
@@ -85,6 +89,11 @@ public class Eth2SigningRequestBody {
   @JsonProperty("fork_info")
   public ForkInfo getForkInfo() {
     return fork_info;
+  }
+
+  @JsonProperty("epoch")
+  public UInt64 getEpoch() {
+    return epoch;
   }
 
   @JsonProperty("block")
@@ -147,8 +156,8 @@ public class Eth2SigningRequestBody {
     return contributionAndProof;
   }
 
-  @JsonProperty("validator_registration_request")
-  public ValidatorRegistrationRequest getValidatorRegistrationRequest() {
-    return validatorRegistrationRequest;
+  @JsonProperty("validator_registration")
+  public ValidatorRegistration getValidatorRegistration() {
+    return validatorRegistration;
   }
 }
